Moved Docking class to its own file.
Made project references point to a less system-specific folder (if this solution folder is at the same level in the directory tree as the SEModAPI solution folder, the references will resolve).
